requirements.txt.txt
aiogram==3.22.0
aiosqlite==0.19.0
aiohttp==3.9.1
python-dotenv==1.0.0
cryptopay-api==0.1.5
requests==2.31.0
pydantic==2.5.0